+++ app/src/main/java/com/jmstudios/redmoon/Changelog.kt
@@ -1,19 +1,28 @@
 /*
  * Copyright (c) 2016  Marien Raat <marienraat@riseup.net>
  * Copyright (c) 2017  Stephen Michel <s@smichel.me>
- * SPDX-License-Identifier: GPL-3.0+
+ * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
 package com.jmstudios.redmoon
 
 import android.app.Activity
 import android.app.Dialog
-import android.support.v7.app.AlertDialog
+import androidx.appcompat.app.AlertDialog
 
 import com.jmstudios.redmoon.model.Config
 
 import java.io.BufferedReader
 import java.io.InputStreamReader
+
+fun showChangelogAuto(activity: Activity): Dialog? {
+    val lastShown = Config.lastChangelogShown
+    Config.lastChangelogShown = BuildConfig.VERSION_CODE
+    return when (lastShown) {
+        0, BuildConfig.VERSION_CODE -> null // First run or already shown
+        else -> showChangelog(activity)
+    }
+}
 
 fun showChangelog(activity: Activity): Dialog? {
     val changelog = StringBuilder()
@@ -31,6 +40,5 @@ fun showChangelog(activity: Activity): Dialog? {
         create()
     }
     dialog.show()
-    Config.lastChangelogShown = BuildConfig.VERSION_CODE
     return dialog
 }