+++ app/src/main/java/com/jmstudios/redmoon/ThemedAppCompatActivity.kt
@@ -1,14 +1,15 @@
 /*
  * Copyright (c) 2016 Marien Raat <marienraat@riseup.net>
  * Copyright (c) 2017  Stephen Michel <s@smichel.me>
- * SPDX-License-Identifier: GPL-3.0+
+ * SPDX-License-Identifier: GPL-3.0-or-later
  */
 package com.jmstudios.redmoon
 
 import android.os.Bundle
-import android.preference.PreferenceFragment
-import android.support.v7.app.AppCompatActivity
 import android.view.MenuItem
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
+import androidx.preference.PreferenceFragmentCompat
 
 import com.jmstudios.redmoon.R
 
@@ -17,28 +18,31 @@ import com.jmstudios.redmoon.util.*
 
 abstract class ThemedAppCompatActivity : AppCompatActivity() {
 
-    protected abstract val fragment: PreferenceFragment
+    protected abstract val fragment: PreferenceFragmentCompat
     protected abstract val tag: String
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        setTheme(Config.activeTheme)
-        setContentView(R.layout.activity_main)
+        super.onCreate(savedInstanceState)
+        setTheme(R.style.AppTheme)
+        setContentView(R.layout.activity_fragment_container)
+
+        val toolbar: Toolbar = findViewById(R.id.toolbar)
+        setSupportActionBar(toolbar)
 
         // Only create and attach a new fragment on the first Activity creation.
         if (savedInstanceState == null) {
             Log.i("onCreate - First creation")
-            fragmentManager.beginTransaction()
-                           .replace(R.id.fragment_container, fragment, tag)
-                           .commit()
+            supportFragmentManager
+                .beginTransaction()
+                .replace(R.id.fragment_container, fragment, tag)
+                .commit()
         }
-
-        super.onCreate(savedInstanceState)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
             android.R.id.home -> {
-                finish()
+                onBackPressed()
                 return true
             }