+++ app/src/main/java/com/jmstudios/redmoon/schedule/LocationUpdateService.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2016  Marien Raat <marienraat@riseup.net>
  * Copyright (c) 2017  Stephen Michel <s@smichel.me>
- * SPDX-License-Identifier: GPL-3.0+
+ * SPDX-License-Identifier: GPL-3.0-or-later
  *
  * This file incorporates work covered by the following copyright and
  * permission notice:
@@ -45,7 +45,7 @@ import java.util.concurrent.TimeUnit
  * location fix, we shut down the service. When we shut down (for any reason),
  * we update the location with the last known location. This way, even if we
  * didn't get a fix before the service was stopped, we might be able to get
- * something more recent than the last time red moon updated location.
+ * something more recent than the last time Blue Light Filter updated location.
  */
 class LocationUpdateService: Service(), LocationListener {