+++ app/src/main/java/com/jmstudios/redmoon/util/Util.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2016  Marien Raat <marienraat@riseup.net>
  * Copyright (c) 2017  Stephen Michel <s@smichel.me>
- * SPDX-License-Identifier: GPL-3.0+
+ * SPDX-License-Identifier: GPL-3.0-or-later
  *
  * This file incorporates work covered by the following copyright and
  * permission notice:
@@ -25,24 +25,27 @@
 package com.jmstudios.redmoon.util
 
 import android.content.Intent
-import android.preference.Preference
-import android.preference.PreferenceFragment
-import android.support.v4.content.ContextCompat
-
+import android.net.Uri
+import android.util.Log
+import androidx.core.content.ContextCompat
+import androidx.preference.Preference
+import androidx.preference.PreferenceFragmentCompat
 import com.jmstudios.redmoon.RedMoonApplication
-import com.jmstudios.redmoon.model.Profile
+import com.jmstudios.redmoon.appextensions.AppExtensionState
+import com.jmstudios.redmoon.appextensions.getContentUri
 import com.jmstudios.redmoon.model.Config
-
+import com.jmstudios.redmoon.model.Profile
+import java.util.*
 import kotlin.reflect.KClass
 
 val appContext = RedMoonApplication.app
 
 var activeProfile: Profile
-    get() = EventBus.getSticky(Profile::class) ?: with (Config) {
-                Profile(color, intensity, dimLevel, lowerBrightness)
-            }
+    get() = EventBus.getSticky(Profile::class) ?: with(Config) {
+        Profile(color, intensity, dimLevel, lowerBrightness)
+    }
     set(value) = value.let {
-        if (it != EventBus.getSticky(Profile::class)) with (Config) {
+        if (it != EventBus.getSticky(Profile::class)) with(Config) {
             val Log = KLogging.logger("Util")
             Log.i("activeProfile set to $it")
             EventBus.postSticky(it)
@@ -56,18 +59,59 @@ var activeProfile: Profile
 var filterIsOn: Boolean = false
     set(value) {
         field = value
+        val appExtensionState = if (field) {
+            AppExtensionState.START
+        } else {
+            AppExtensionState.STOP
+        }
+        val uri: Uri = getContentUri(appExtensionState.id)
+        appContext.contentResolver.insert(uri, null)
+        Log.d("AppExtensionTest", "filterIsOn $uri ${appExtensionState.id}")
         Config.filterIsOn = value
     }
 
+
+fun inActivePeriod(Log: KLog? = null): Boolean {
+    val now = Calendar.getInstance()
+
+    val onTime = Config.scheduledStartTime
+    val onHour = onTime.substringBefore(':').toInt()
+    val onMinute = onTime.substringAfter(':').toInt()
+    val on = Calendar.getInstance().apply {
+        set(Calendar.HOUR_OF_DAY, onHour)
+        set(Calendar.MINUTE, onMinute)
+        if (after(now)) {
+            add(Calendar.DATE, -1)
+        }
+    }
+
+    val offTime = Config.scheduledStopTime
+    val offHour = offTime.substringBefore(':').toInt()
+    val offMinute = offTime.substringAfter(':').toInt()
+    val off = Calendar.getInstance().apply {
+        set(Calendar.HOUR_OF_DAY, offHour)
+        set(Calendar.MINUTE, offMinute)
+        while (before(on)) {
+            add(Calendar.DATE, 1)
+        }
+    }
+
+    Log?.d("Start: $onTime, stop: $offTime")
+    Log?.d("On DAY_OF_MONTH: ${on.get(Calendar.DAY_OF_MONTH)}")
+    Log?.d("Off DAY_OF_MONTH: ${off.get(Calendar.DAY_OF_MONTH)}")
+
+    return now.after(on) && now.before(off)
+}
+
 fun getString(resId: Int): String = appContext.getString(resId)
-fun getColor (resId: Int): Int = ContextCompat.getColor(appContext, resId)
+fun getColor(resId: Int): Int = ContextCompat.getColor(appContext, resId)
 
 fun atLeastAPI(api: Int): Boolean = android.os.Build.VERSION.SDK_INT >= api
-fun belowAPI  (api: Int): Boolean = android.os.Build.VERSION.SDK_INT <  api
+fun belowAPI(api: Int): Boolean = android.os.Build.VERSION.SDK_INT < api
 
 fun intent() = Intent()
-fun <T: Any>intent(kc: KClass<T>) = Intent(appContext, kc.java)
+fun <T : Any> intent(kc: KClass<T>) = Intent(appContext, kc.java)
 
-fun PreferenceFragment.pref(resId: Int): Preference {
+fun PreferenceFragmentCompat.pref(resId: Int): Preference? {
     return preferenceScreen.findPreference(getString(resId))
 }